# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Live Music Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface User {
  id: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  isAdmin: boolean;
  createdAt: string;
}

export interface UserPreferences {
  id: number;
  userId: number;
  cities: string[];
  zipCodes: string[];
  updatedAt: string;
}

export interface UpdateUserPreferencesBody {
  cities: string[];
  zipCodes: string[];
}

export interface Venue {
  id: number;
  name: string;
  city: string;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zipCode?: string | null;
  websiteUrl: string;
  /** @nullable */
  scrapeUrl?: string | null;
  isActive: boolean;
  /** @nullable */
  lastScrapedAt?: string | null;
  createdAt: string;
}

export interface CreateVenueBody {
  name: string;
  city: string;
  state?: string;
  zipCode?: string;
  websiteUrl: string;
  scrapeUrl?: string;
  isActive?: boolean;
}

export interface UpdateVenueBody {
  name?: string;
  city?: string;
  state?: string;
  zipCode?: string;
  websiteUrl?: string;
  scrapeUrl?: string;
  isActive?: boolean;
}

export interface Show {
  id: number;
  venueId: number;
  title: string;
  /** @nullable */
  artist?: string | null;
  /** @nullable */
  description?: string | null;
  showDate: string;
  /** @nullable */
  doorsTime?: string | null;
  /** @nullable */
  showTime?: string | null;
  /** @nullable */
  ticketUrl?: string | null;
  /** @nullable */
  ticketPrice?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  sourceUrl?: string | null;
  createdAt: string;
}

export interface AttendeeInfo {
  userId: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  boughtTickets: boolean;
}

export interface ShowWithDetails {
  id: number;
  venueId: number;
  title: string;
  /** @nullable */
  artist?: string | null;
  /** @nullable */
  description?: string | null;
  showDate: string;
  /** @nullable */
  doorsTime?: string | null;
  /** @nullable */
  showTime?: string | null;
  /** @nullable */
  ticketUrl?: string | null;
  /** @nullable */
  ticketPrice?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  sourceUrl?: string | null;
  createdAt: string;
  venue: Venue;
  attendeeCount: number;
  currentUserAttending: boolean;
  currentUserBoughtTickets: boolean;
  friendsAttending: AttendeeInfo[];
}

export interface Attendance {
  id: number;
  userId: number;
  showId: number;
  boughtTickets: boolean;
  createdAt: string;
}

export interface AttendanceBody {
  boughtTickets: boolean;
}

export interface FriendInfo {
  friendshipId: number;
  userId: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  upcomingShowsCount: number;
}

export type FriendRequestStatus =
  (typeof FriendRequestStatus)[keyof typeof FriendRequestStatus];

export const FriendRequestStatus = {
  pending: "pending",
  accepted: "accepted",
  declined: "declined",
} as const;

export interface FriendRequest {
  id: number;
  fromUserId: number;
  toUserId: number;
  fromUsername: string;
  toUsername: string;
  status: FriendRequestStatus;
  createdAt: string;
}

export interface SendFriendRequestBody {
  toUserId: number;
}

export interface FriendActivity {
  showId: number;
  showTitle: string;
  showDate: string;
  venueName: string;
  venueCity: string;
  /** @nullable */
  ticketUrl?: string | null;
  friends: AttendeeInfo[];
}

export interface UserSearchResult {
  id: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  isFriend: boolean;
  hasPendingRequest: boolean;
}

export interface ScrapeResult {
  venueId: number;
  venueName: string;
  showsFound: number;
  showsAdded: number;
  errors: string[];
}

export interface ScrapeAllResult {
  totalShowsFound: number;
  totalShowsAdded: number;
  venuesScraped: number;
  results: ScrapeResult[];
}

export type ListVenuesParams = {
  active?: boolean;
};

export type ListShowsParams = {
  city?: string;
  zipCode?: string;
  startDate?: string;
  endDate?: string;
  venueId?: number;
};

export type SearchUsersParams = {
  q: string;
};

export type ExportShowsParams = {
  city?: string;
  zipCode?: string;
  startDate?: string;
  endDate?: string;
};