# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Live Music Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /friends/requests:
    get:
      operationId: listFriendRequests
      tags: [friends]
      summary: List pending friend requests
      responses:
        "200":
          description: Friend requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FriendRequest"
    post:
      operationId: sendFriendRequest
      tags: [friends]
      summary: Send a friend request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendFriendRequestBody"
      responses:
        "201":
          description: Friend request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequest"
  /friends/requests/{id}/accept:
    post:
      operationId: acceptFriendRequest
      tags: [friends]
      summary: Accept a friend request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /friends/requests/{id}/decline:
    post:
      operationId: declineFriendRequest
      tags: [friends]
      summary: Decline a friend request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Declined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /friends/{id}:
    delete:
      operationId: removeFriend
      tags: [friends]
      summary: Remove a friend
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
  /export/shows:
    get:
      operationId: exportShows
      tags: [export]
      summary: Export shows as CSV
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: zipCode
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    FriendRequest:
      type: object
      properties:
        id:
          type: integer
        fromUserId:
          type: integer
        toUserId:
          type: integer
        fromUsername:
          type: string
        toUsername:
          type: string
        status:
          type: string
          enum: [pending, accepted, declined]
        createdAt:
          type: string
      required:
        - id
        - fromUserId
        - toUserId
        - fromUsername
        - toUsername
        - status
        - createdAt
    SendFriendRequestBody:
      type: object
      properties:
        toUserId:
          type: integer
      required:
        - toUserId
...