# lib/api-zod/src/generated/types/friendRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Live Music Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { FriendRequestStatus } from "./friendRequestStatus";

export interface FriendRequest {
  id: number;
  fromUserId: number;
  toUserId: number;
  fromUsername: string;
  toUsername: string;
  status: FriendRequestStatus;
  createdAt: string;
}